+++ routes/__init__.py
@@ -1,2 +1,3 @@
 from .home import home_bp
 from .register import register_bp
+from .login import login_pb
