+++ templates/index.html
@@ -19,7 +19,7 @@
             By leveraging data from various sources,<br>
             we can gain insights into behavioral health patterns and improve healthcare practices in the community.</p><br><br>
 
-        <a href="{{ url_for('login') }}"><button>Get Started</button></a>
+        <a href="{{ url_for('login.login') }}"><button>Get Started</button></a>
     </div>
 </body>
 </html>
