+++ templates/register.html
@@ -24,5 +24,5 @@
 
     <button type="submit">Register</button>
 </form>
-<p>Already have an account? <a href="{{ url_for('login') }}">Login here</a></p>
+<p>Already have an account? <a href="{{ url_for('login.login') }}">Login here</a></p>
 {% endblock %}
